Add CLI option to read `compile_commands.json` to find files (#54)
* Add --read-json-db argument to hooks

* Add support for compile_commands.json for lizard hook

* Minor code refactoring

* Add shellcheck pre-commit hook

* Update pre-commit hooks and update CHANGELOG

* Update README

+++ tests/run_tests.sh
@@ -12,29 +12,29 @@ function run_hook()
     gen_compiledb=$4
     shift 4
 
-    mkdir -p $build_dir
-    rm -rf $build_dir/*
+    mkdir -p "$build_dir"
+    rm -rf "${build_dir:?}"/*
 
-    if [ $gen_compiledb -eq 1 ]; then
-        $hook -S $src_dir -B $build_dir "$@"
+    if [ "$gen_compiledb" -eq 1 ]; then
+        $hook -S "$src_dir" -B "$build_dir" "$@"
     else
         $hook "$@"
     fi
     ret=$?
 
-    if [ $expect_success -eq 1 ]; then
+    if [ "$expect_success" -eq 1 ]; then
         if [ $ret -ne 0 ]; then
             echo "Hook $hook failed on source directory $src_dir with exit code $ret!"
             exit 1
         fi
     else
-        if [ $ret -eq 0 ]; then
+        if [ "$ret" -eq 0 ]; then
             echo "Hook $hook unexpectedly succeeded on source directory $src_dir with exit code $ret!"
             exit 1
         fi
     fi
 
-    if [ $gen_compiledb -eq 1 ]; then
+    if [ "$gen_compiledb" -eq 1 ]; then
         if [ ! -f $build_dir/compile_commands.json ]; then
             echo "Hook $hook failed to generate the compile database for $src_dir!"
             exit 1
@@ -52,8 +52,8 @@ run_hook cmake-pc-include-what-you-use-hook tests/cmake_good 1 1 tests/cmake_goo
 run_hook cmake-pc-lizard-hook tests/cmake_good 1 0 tests/cmake_good/good.cpp
 
 # Test multiple build directory options
-mkdir -p $build_dir
-rm -rf $build_dir/*
+mkdir -p "$build_dir"
+rm -rf "${build_dir:?}"/*
 cmake -B other_build -S tests/cmake_good
 cmake-pc-cppcheck-hook -S tests/cmake_good -B build -B other_build tests/cmake_good/good.cpp
 if [ ! -f other_build/compile_commands.json ]; then
@@ -74,7 +74,6 @@ run_hook cmake-pc-lizard-hook tests/cmake_bad 0 0 -C4 tests/cmake_bad/bad.cpp
 
 export CC=clang CXX=clang++
 
-
 if [ -z "$GITHUB_SHA" ]; then
     LATEST_SHA=$(git rev-parse HEAD)
 else
@@ -91,32 +90,32 @@ call_sed()
 }
 
 
-pushd tests/cmake_good
+pushd tests/cmake_good || exit
 git init \
     && git config user.name 'Test' \
     && git config user.email 'test@test.com' \
     && call_sed -i "/rev: /c\\    rev: ${LATEST_SHA}" .pre-commit-config.yaml \
-    && git add *.txt *.cpp .pre-commit*.yaml \
+    && git add ./*.txt ./*.cpp .pre-commit*.yaml \
     && git commit -m 'Initial commit' \
     && pre-commit run --all-files && success=1 || success=0
 rm -rf .git
 
-if [ $success -eq 0 ]; then
+if [ "$success" -eq 0 ]; then
     exit 1
 fi
-popd
+popd || exit
 
-pushd tests/cmake_bad
+pushd tests/cmake_bad || exit
 git init \
     && git config user.name 'Test' \
     && git config user.email 'test@test.com' \
     && call_sed -i "/rev: /c\\    rev: ${LATEST_SHA}" .pre-commit-config.yaml \
-    && git add *.txt *.cpp .pre-commit*.yaml \
+    && git add ./*.txt ./*.cpp .pre-commit*.yaml \
     && git commit -m 'Initial commit' \
     && pre-commit run --all-files && success=0 || success=1
 rm -rf .git
 
-if [ $success -eq 0 ]; then
+if [ "$success" -eq 0 ]; then
     exit 1
 fi
-popd
+popd || exit
